ENH Extracting logic for canX from Company and Employee into DataObject Extension
