+++ code/Company.php
@@ -13,8 +13,6 @@ use SilverStripe\Versioned\RecursivePublishable;
 use SilverStripe\Versioned\Versioned;
 use RelationFieldsTestPage;
 use GridFieldTestPage;
-use SilverStripe\Security\Permission;
-use SilverStripe\Security\PermissionProvider;
 
 /**
  *
@@ -33,7 +31,7 @@ use SilverStripe\Security\PermissionProvider;
  * @mixin Versioned
  * @mixin RecursivePublishable
  */
-class Company extends DataObject implements PermissionProvider
+class Company extends DataObject
 {
     private static $table_name = 'Company';
 
@@ -371,42 +369,6 @@ class Company extends DataObject implements PermissionProvider
     public function scaffoldSearchField()
     {
         return DropdownField::create('CompanyID', 'Company', self::get()->map())->setEmptyString('');
-    }
-
-    public function providePermissions()
-    {
-        return [
-            'COMPANY_EDIT' => [
-                'name' => _t(
-                    __CLASS__ . '.EditPermissionLabel',
-                    'Edit a company'
-                ),
-                'category' => _t(
-                    __CLASS__ . '.Category',
-                    'Company'
-                ),
-            ],
-        ];
-    }
-
-    public function canView($member = null) 
-    {
-        return Permission::check('COMPANY_EDIT', 'any', $member);
-    }
-
-    public function canEdit($member = null) 
-    {
-        return Permission::check('COMPANY_EDIT', 'any', $member);
-    }
-
-    public function canDelete($member = null) 
-    {
-        return Permission::check('COMPANY_EDIT', 'any', $member);
-    }
-
-    public function canCreate($member = null, $context = []) 
-    {
-        return Permission::check('COMPANY_EDIT', 'any', $member);
     }
 
 }
